Moved tile, extent, projected_extent, temporal_projected_extent, and keys to geoypyspark/geotrellis
