+++ geopyspark/tests/schemas/extent_test.py
@@ -1,9 +1,7 @@
-#!/bin/env python3
-
 from pyspark import SparkConf, SparkContext, RDD
 from pyspark.serializers import Serializer, FramedSerializer, AutoBatchedSerializer
 from py4j.java_gateway import java_import
-from geopyspark.extent import Extent
+from geopyspark.geotrellis.extent import Extent
 from geopyspark.avroserializer import AvroSerializer
 from geopyspark.avroregistry import AvroRegistry
