+++ geopyspark/tests/schemas/keys_test.py
@@ -1,9 +1,7 @@
-#!/user/bin/env python3
-
 from pyspark import SparkConf, SparkContext, RDD
 from pyspark.serializers import Serializer, FramedSerializer, AutoBatchedSerializer
 from py4j.java_gateway import java_import
-from geopyspark.keys import SpatialKey, SpaceTimeKey
+from geopyspark.geotrellis.keys import SpatialKey, SpaceTimeKey
 from geopyspark.avroserializer import AvroSerializer
 from geopyspark.avroregistry import AvroRegistry
