+++ geopyspark/tests/schemas/projected_extent_test.py
@@ -1,11 +1,9 @@
-#!/bin/env python3
-
 from pyspark import SparkConf, SparkContext, RDD
 from pyspark.serializers import Serializer, FramedSerializer, AutoBatchedSerializer
 from py4j.java_gateway import java_import
 from geopyspark.avroserializer import AvroSerializer
-from geopyspark.extent import Extent
-from geopyspark.projected_extent import ProjectedExtent
+from geopyspark.geotrellis.extent import Extent
+from geopyspark.geotrellis.projected_extent import ProjectedExtent
 from geopyspark.avroregistry import AvroRegistry
 
 import unittest
