+++ geopyspark/tests/schemas/tile_test.py
@@ -1,11 +1,9 @@
-#!/bin/env python3
-
 from pyspark import SparkConf, SparkContext, RDD
 from pyspark.serializers import Serializer, FramedSerializer, AutoBatchedSerializer
 from py4j.java_gateway import java_import
 from geopyspark.avroserializer import AvroSerializer
 from geopyspark.avroregistry import AvroRegistry
-from geopyspark.tile import TileArray
+from geopyspark.geotrellis.tile import TileArray
 
 import numpy as np
 import unittest
